requirements.txt
Flask==3.0.3
Pillow==10.4.0
torch==2.3.0
transformers==4.44.2
